requirements.txt
biopython>=1.74
numpy>=1.16.4
